The primary interface's state is checked first and if it is up, it will check for internet connectivity with a single ping to google.com. If the ping is successful, the consecutive_failures counter is reset to 0, and the primary_iface_failed flag is set to false. If the ping is not successful, the consecutive_failures counter is incremented. If the consecutive_failures counter reaches the PING_TIMEOUT threshold, the primary interface is considered down, the primary_iface_fail_time is set to the current time, the primary_iface_failed flag is set to true, the primary interface is brought down and the secondary interface is brought up.
If the primary interface is down, the program will check whether the primary_iface_failed flag is set to true. If it is, it will check the time elapsed since the primary interface failed, if the time elapsed is greater than or equal to the PRIMARY_IFACE_RECOVERY_TIME, the primary interface is considered recovered and the secondary interface is brought down and the primary interface is brought up.

The sleep function is called at the end of each iteration, to slow down the rate of the loop, this is set to 1Hz.